move storage construct to be internal to backend-storage
