+++ packages/backend-storage/package.json
@@ -18,8 +18,8 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
+    "@aws-amplify/backend-output-schemas": "^0.4.0",
     "@aws-amplify/backend-output-storage": "^0.2.6",
-    "@aws-amplify/storage-construct-alpha": "^0.2.3",
     "@aws-amplify/plugin-types": "^0.5.0"
   },
   "devDependencies": {
