+++ packages/backend-storage/src/index.ts
@@ -1,1 +1,2 @@
 export * from './factory.js';
+export { StorageResources, AmplifyStorageProps } from './construct.js';
