+++ packages/backend-storage/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": { "rootDir": "src", "outDir": "lib" },
   "references": [
     { "path": "../backend-output-storage" },
-    { "path": "../storage-construct" },
     { "path": "../plugin-types" },
     { "path": "../backend-platform-test-stubs" },
     { "path": "../platform-core" }
