--- packages/storage-construct/.npmignore
@@ -1,14 +1,0 @@
-# Be very careful editing this file. It is crafted to work around [this issue](https://github.com/npm/npm/issues/4479)
-
-# First ignore everything
-**/*
-
-# Then add back in transpiled js and ts declaration files
-!lib/**/*.js
-!lib/**/*.d.ts
-
-# Then ignore test js and ts declaration files
-*.test.js
-*.test.d.ts
-
-# This leaves us with including only js and ts declaration files of functional code
