--- packages/storage-construct/API.md
@@ -1,33 +1,0 @@
-## API Report File for "@aws-amplify/storage-construct-alpha"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { BackendOutputStorageStrategy } from '@aws-amplify/plugin-types';
-import { Construct } from 'constructs';
-import { IBucket } from 'aws-cdk-lib/aws-s3';
-import { ResourceProvider } from '@aws-amplify/plugin-types';
-import { StorageOutput } from '@aws-amplify/backend-output-schemas';
-
-// @public
-export class AmplifyStorage extends Construct implements ResourceProvider<StorageResources> {
-    constructor(scope: Construct, id: string, props: AmplifyStorageProps);
-    // (undocumented)
-    readonly resources: StorageResources;
-}
-
-// @public (undocumented)
-export type AmplifyStorageProps = {
-    versioned?: boolean;
-    outputStorageStrategy?: BackendOutputStorageStrategy<StorageOutput>;
-};
-
-// @public (undocumented)
-export type StorageResources = {
-    bucket: IBucket;
-};
-
-// (No @packageDocumentation comment for this package)
-
-```
