--- packages/storage-construct/CHANGELOG.md
@@ -1,173 +1,0 @@
-# @aws-amplify/storage-construct-alpha
-
-## 0.2.3
-
-### Patch Changes
-
-- 5ed51cbd5: Upgrade aws-cdk to 2.110.1
-- Updated dependencies [5ed51cbd5]
-  - @aws-amplify/backend-output-storage@0.2.6
-
-## 0.2.2
-
-### Patch Changes
-
-- Updated dependencies [07b0dfc9f]
-  - @aws-amplify/backend-output-schemas@0.4.0
-  - @aws-amplify/backend-output-storage@0.2.2
-
-## 0.2.1
-
-### Patch Changes
-
-- Updated dependencies [71a63a16]
-  - @aws-amplify/backend-output-schemas@0.3.0
-  - @aws-amplify/backend-output-storage@0.2.1
-
-## 0.2.0
-
-### Minor Changes
-
-- 47456c26: Store attribution metadata in stack descriptions
-
-### Patch Changes
-
-- 0398b8e1: Bump graphql construct to 0.9.0 and remove some interface cruft
-- b2b0c2da: force version bump
-- 7296e9d9: Initial publish
-- 3bda96ff: update methods to use arrow notation
-- 7103735b: cdk lib dependency declaration
-- 36d93e46: add license to package.json
-- 8f99476e: chore: upgrade aws-cdk to 2.103.0
-- 47456c26: Remove ESM features from construct dependency packages and make corresponding updates in consumer packages
-- f75fa531: Refactor OutputStorageStrategy into stateless shared dependency
-- 59f5ea24: chore: upgrade aws-cdk to 2.100.0
-- Updated dependencies [47456c26]
-- Updated dependencies [ac3df080]
-- Updated dependencies [0398b8e1]
-- Updated dependencies [b2b0c2da]
-- Updated dependencies [7296e9d9]
-- Updated dependencies [53779253]
-- Updated dependencies [395c8f0d]
-- Updated dependencies [ce008a2c]
-- Updated dependencies [36d93e46]
-- Updated dependencies [8f99476e]
-- Updated dependencies [47456c26]
-- Updated dependencies [b4f82717]
-- Updated dependencies [05f97b26]
-- Updated dependencies [2525b582]
-- Updated dependencies [f75fa531]
-- Updated dependencies [f6618771]
-- Updated dependencies [f201c94a]
-- Updated dependencies [883d9da7]
-- Updated dependencies [59f5ea24]
-  - @aws-amplify/backend-output-storage@0.2.0
-  - @aws-amplify/backend-output-schemas@0.2.0
-
-## 0.2.0-alpha.11
-
-### Minor Changes
-
-- 47456c26: Store attribution metadata in stack descriptions
-
-### Patch Changes
-
-- 47456c26: Remove ESM features from construct dependency packages and make corresponding updates in consumer packages
-- Updated dependencies [47456c26]
-- Updated dependencies [47456c26]
-  - @aws-amplify/backend-output-storage@0.2.0-alpha.6
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.8
-
-## 0.1.1-alpha.10
-
-### Patch Changes
-
-- 8f99476e: chore: upgrade aws-cdk to 2.103.0
-- Updated dependencies [8f99476e]
-  - @aws-amplify/backend-output-storage@0.2.0-alpha.5
-
-## 0.1.1-alpha.9
-
-### Patch Changes
-
-- Updated dependencies [883d9da7]
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.7
-  - @aws-amplify/backend-output-storage@0.2.0-alpha.4
-
-## 0.1.1-alpha.8
-
-### Patch Changes
-
-- 59f5ea24: chore: upgrade aws-cdk to 2.100.0
-- Updated dependencies [59f5ea24]
-  - @aws-amplify/backend-output-storage@0.1.1-alpha.3
-
-## 0.1.1-alpha.7
-
-### Patch Changes
-
-- 7103735b: cdk lib dependency declaration
-
-## 0.1.1-alpha.6
-
-### Patch Changes
-
-- 36d93e46: add license to package.json
-- Updated dependencies [36d93e46]
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.5
-  - @aws-amplify/backend-output-storage@0.1.1-alpha.2
-
-## 0.1.1-alpha.5
-
-### Patch Changes
-
-- 0398b8e1: Bump graphql construct to 0.9.0 and remove some interface cruft
-- Updated dependencies [0398b8e1]
-- Updated dependencies [f6618771]
-  - @aws-amplify/backend-output-storage@0.1.1-alpha.1
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.4
-
-## 0.1.1-alpha.4
-
-### Patch Changes
-
-- f75fa531: Refactor OutputStorageStrategy into stateless shared dependency
-- Updated dependencies [ac3df080]
-- Updated dependencies [53779253]
-- Updated dependencies [1dada824]
-- Updated dependencies [b4f82717]
-- Updated dependencies [05f97b26]
-- Updated dependencies [f75fa531]
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.3
-  - @aws-amplify/backend-output-storage@0.1.1-alpha.0
-
-## 0.1.1-alpha.3
-
-### Patch Changes
-
-- Updated dependencies [ce008a2]
-- Updated dependencies [f201c94]
-  - @aws-amplify/backend-output-schemas@0.2.0-alpha.2
-
-## 0.1.1-alpha.2
-
-### Patch Changes
-
-- b2b0c2d: force version bump
-- Updated dependencies [b2b0c2d]
-- Updated dependencies [395c8f0]
-  - @aws-amplify/backend-output-schemas@0.1.1-alpha.1
-
-## 0.1.1-alpha.1
-
-### Patch Changes
-
-- 3bda96f: update methods to use arrow notation
-
-## 0.1.1-alpha.0
-
-### Patch Changes
-
-- 7296e9d: Initial publish
-- Updated dependencies [7296e9d]
-  - @aws-amplify/backend-output-schemas@0.1.1-alpha.0
