--- packages/storage-construct/api-extractor.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../api-extractor.base.json"
-}
