--- packages/storage-construct/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@aws-amplify/storage-construct-alpha",
-  "version": "0.2.3",
-  "type": "module",
-  "publishConfig": {
-    "access": "public"
-  },
-  "exports": {
-    ".": {
-      "types": "./lib/index.d.ts",
-      "import": "./lib/index.js",
-      "require": "./lib/index.js"
-    }
-  },
-  "types": "lib/index.d.ts",
-  "scripts": {
-    "update:api": "api-extractor run --local"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@aws-amplify/backend-output-schemas": "^0.4.0",
-    "@aws-amplify/backend-output-storage": "^0.2.6"
-  },
-  "devDependencies": {
-    "@aws-amplify/plugin-types": "^0.5.0"
-  },
-  "peerDependencies": {
-    "aws-cdk-lib": "^2.110.1",
-    "constructs": "^10.0.0"
-  }
-}
