--- packages/storage-construct/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": { "rootDir": "src", "outDir": "lib" },
-  "references": [
-    { "path": "../backend-output-schemas" },
-    { "path": "../backend-output-storage" },
-    { "path": "../plugin-types" }
-  ]
-}
